refactor: 상품 가격 조회 서비스의 인터페이스를 추출하여 상품에 대한 의존성을 상품 패키지 내부로 이전
