+++ src/test/java/kitchenpos/menus/application/FakeProductPriceReader.java
@@ -1,0 +1,28 @@
+package kitchenpos.menus.application;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.exception.ProductNotFoundException;
+import kitchenpos.reader.ProductPriceReader;
+
+public class FakeProductPriceReader implements ProductPriceReader {
+
+    private final ProductRepository productRepository;
+
+    public FakeProductPriceReader(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Override
+    public BigDecimal getProductPriceById(UUID productId) {
+        Product product = findProductById(productId);
+        return product.getPriceValue();
+    }
+
+    private Product findProductById(UUID productId) {
+        return productRepository.findById(productId)
+            .orElseThrow(() -> new ProductNotFoundException("ID 에 해당하는 상품을 찾을 수 없습니다."));
+    }
+}
